refactor rename examples.ts to example-option.ts
