+++ packages/create-waku/src/helpers/example-option.ts
